Fix: input text color for light mode
I just discovered there is a css variable `--color-character-text` specifically meant for the text color used on the the item color background. This is what should be used to match the Wanikani default and Breeze Dark 2 makes sure to override this value so it will still work with that style.