Fix: return to dashboard when lessons still left & bump min wkof ver
When the user chose to continue with lessons, it would return to the dashboard anyway because the `currentLessonQueue` had been emptied before visit. This was due to an accidental loop caused by page visits. A flag that signals a modified start after a quiz completes fixes this. Also bumped the minimum wkof version to the one that supports the new lesson url format.